TIDY: Usage drivers get their own module
